Add preprocessing of models at integral-comparison detection method
